Unbreak build on OpenBSD and other OS that aren't linux, windows, or macos
The build.rd script tries to find an android toolchain but fails on
other OS, such as the BSDs, that do not provide one.  This change
simply marks them as "unknown" and fails to build with an unknown
android toolchain, but succeeds to build the regular crate on the any
other OS natively.  The fix is not specific to OpenBSD as it unbreaks
all OS that are not linux, windows, or macos.