[Test] Move serde strucutre tests out of the library.
These tests require `serde_test` as a dependency which automatically
introduces a copy of `serde`, with enabled default features, into the
dependecy tree.

As cargo forbids optional dev-dependencies there was no chance to keep
it in this list if we want to allow to use the `serde` feature in
`no_std` mode. It was therefore  necessary to move the `serde_test`
dependency from the list of dev-dependencies into the list of normal
dependencies.

**Attention**

This will require to specify the features `serde_test` and `serde` to
run the new integration test.