Pass full public key to hash function instead of its coordinates
The `new_with_hash` constructor of `SharedSecret` passes a `null` hash
function to the underlying C code. This means what we are getting passed
is the literal result of the DH operation which is a PublicKey.

We can model this more closely by making the type of the closure an
actual PublicKey. As a nice side-effect, we can avoid copying the return
value.