Enable build with global-context-less-secure
Currently this command fails:

 cargo build --no-default-features --features=global-context-less-secure

All features should be able to be built individually, this is currently
not the case with `global-context-less-secure`.

Enable `std` if `global-context-less-secure` is enabled (because `Once`
only comes from `std`, not available in `core`).

Add `global-context-less-secure` to the features test array in
`contrib/test.sh`.

With this applied the build command above runs successfully.