Add bitcoin_hashes features
We re-export `bitcoin_hashes` so users do not have to depend directly on
it. This is Rust best practice as makes keeping versions in sync easier.

Currently, however, it is not possible to enable different features in
`bitcoin_hashes` and our feature enables it with no default features -
this makes the re-export basically useless.

Add a bunch of features that can be used to turn on the various
`bitcoin_hashes` features.