Use clang instead of clang-9
The current version of clang is 14, there is no obvious reason why we
use clang-9 (as far as I can tell on my local machine).

Use `clang` instead of `clang-9` so that the latest version is used by
default. This effects the version installed by CI as well as the
version used to run commands in `test.sh`.