Duplicate impl_array_newtype macro
Currently we define the `impl_array_newtype` in `secp256k1-sys` and use
it both there and in `rust-secp256k1`, this is conceptually incorrect
because the inner array in `rust-secp256k1` is fully controlled by the
Rust language where as the inner array for types in `secp256k1-sys` is
passed across the FFI boundary and as such there are no guarantees to
layout of the array across versions of `libsecp256k1`.

Duplicate the `impl_array_newtype` macro in
`rust-secp256k1/src/macros.rs` and export the new version instead of the
one from `sepc256k1-sys`. No other changes to the macro so this patch
introduces no logic changes.