Remove derives for ffi wrapped types
Currently we are deriving various traits that call down to the traits
implemented in secp265k1-sys that in turn call down to implementations
on the inner byte array. Since we cannot guarantee the inner byte arrays
in secp256k1-sys this results in trait implementations that may not be
stable across versions of the library.

For `XOnlyPublicKey`, `KeyPair`, and `ecdsa::Signature` manually
implement `PartialEq`, `Eq`, `PartialOrd`, `Ord`, and `Hash` by first
converting the type into a form that is guaranteed to be stable across
library versions e.g., by serializing it to a known format.