Implement fast unstable comparisons
We recently improved the comparison implementations to be stable across
library versions by first serializing the data types before comparing
them instead of just comparing the inner bytes. This had the downside
that the comparisons are now slower.

Add public methods `cmp_fast_unstable` and `eq_fast_unstable` to each
of the types that wraps an FFI type.