Guard code with bitcoin_hashes not bitcoin-hashes
We (I) recently added a feature alias "bitcoin-hashes" to try to make
the use of the library more ergonomic, in doing so we also used
"bitcoin-hashes" to feature guard code - this has negative consequences
because it leads to a situation where a devs typical workflow of
enabling a feature by using its dependency name breaks because one can
enable "bitcoin_hashes" and not get the feature gated code. We already
have a whole mess because of "std" so the only people using
"bitcoin_hashes" should be those doing "no-std" builds but still this
was a regression.

Originally the goal was to shield users from the historical mistake of
using the underscore that cannot be fixed. This is still an admirable
goal, hence in docs and example code keep using "bitcoin-hashes" as the
feature. And since we re-export `bitcoin_hashes` both here in secp and
in bitcoin new users will hopefully just use `secp256k1::hashes` and
slowly start to forget about `bitcoin_hashes` :)