backport: fix rand-std feature for old cargo versions
There is a bugfix https://github.com/rust-lang/cargo/pull/8395
which appears in cargo 1.46, after our MSRV of 1.41. This means
that under 1.41, users of the rand-std feature may nondeterministically
see compilation fail.

In particular, users of rust-bitcoin 0.29 seem to have started seeing
failures starting on 2023-01-12, where the 'rand-std' feature would be
enabled in this library but 'rand' was not.

This commit simply makes the 'rand' dependency explicit, to avoid the
nondeterministic bug in cargo's feature resolver.