Patch the bitcoin_hashes dependency
Currently we have a hole in our dependency graph that is making it
impossible to patch `bitcoin_hashes` with breaking changes.

  bitcoin -> secp -> hashes
          -> hashes

Make `secp256k1` depend on the first commit that imported
`bitcoin_hashes` into `rust-bitcoin`. No code changes required.

With this applied, when patching `hashes` with a breaking change,
we will be able to build the code still.