Add changelog entry for removal of ONE_KEY
In the 0.25.0 release we removed the `ONE_KEY` const but did not mention
it in the release notes. This makes upgrade harder than it should be.

In hindsight we should have deprecated the const then removed it.

Add a changelog entry to the 0.25.0 release mentioning the removal and
the work around.