Use flag when building secp sys
We currently use the `cc::Build::flag_if_supported` function to pass in
`-Wno-unused-function` however when building for wasm the flag is
ignored.

Using plain old `flag` seems to work, found by trial and error, I did no
further investigation as to why it works and the other does not.

Resolves: #434