Stop implementing ThirtyTwoByteHash
The implementations of `ThirtyTwoByteHash` for types from the `hashes`
crate are problematic during upgrades because both `bitcoin` and
`secp256k1` depend on `hashes` and when the versions of `hashes` get out
of sync usage of the trait breaks.

In order to resolve this issue stop implementing `ThirtyTwoByteHash` for
the `hashes` types, this means also that we cannot use the
`hashes::Hash` trait bound when implementing `from_hashed_data` for
`SecretKey` or `Message` (because of the bound on `From`). We can still
provide the `from_hashed_data` helper function but we now restrict it to
only hashing with `sha256` instead of a generic hasher. Users can still
create the `Message` and `SecretKey` types from other hashes but to do
so they must do it manually, this is documented in the update rustdocs
for each function.