Test wasm with pinned dependencies
So far testing was was done unpinned however that risked randomly
breaking builds whenever a dependency changes by accident. There was a
note explicitly acknowledging lack of pinning but with no explanation
why that was the case. So it might be because the author didn't want to
bother and pinning can be implemented.