context: whitelist new compiler warning
The compiler doesn't like shared references to static mut variables,
because it is worried that we'll share a reference while a mutation
happens which is UB.

We are only sharing references after using Once to do a one-time
initialization, so this is fine, but in recent versions of the compiler
there is an encapsulated form of this pattern called `OnceLock` and a
variant called `LazyLock`.

Neither of these are available on our MSRV so just whitelist the lint.