musig: a couple small improvements of byte array APIs
I decided not to rename the `serialize` functions to `to_byte_array`.
Maybe we should do that, but we use the name `serialize` *all over* this
library so we should do it in a separate PR that changes everything.
Along the way we'll have to decide what to call the methods that produce
e.g. a SerializedSignature; this is conceptually a byte array but it's
actually not one.