musig: add a bunch of unit tests
I asked Claude to create an initial set of unit tests, which it did. I
then manually cleaned up a lot of its repeated logic (though not all of
it, as you can tell) and added a whole bunch more failure cases. For
example it did not bother trying to repeat or swap keys/nonces.

When it generated the tests, the empty-pubkey-list bug (fixed in the
previous commit) was still present. It failed to find it, I think
because it was reading the code looking for panics to trigger, and there
wasn't one. In future I will try giving it only the API, and try telling
it to be more adversarial. We'll see.