ellswift: remove unused `data` argument from `shared_secret`
We take an optional `data` argument to `shared_secret`, but the upstream
function we call never uses it. However, this argument *does* use the
CPtr impl on &[T] in order to obtain a pointer to pass across the FFI
boundary. This impl is very dangerous, and its use here is sound only
because the resulting pointer is never used.

See
https://github.com/rust-bitcoin/rust-secp256k1/pull/627#discussion_r2172910653