secp256k1-sys: bump version to 0.8.2
This will break CI even further than it's already broken (due to unpinned
dependencies and compilers), because we have a check that the symbol names
in libsecp256k1 match the version number in Cargo.toml.

But updating these names would be a breaking change. In modern versions of
rust-secp we have fixed this by not putting patch version numbers into the
symbol names.