Update `hashes` to `0.16.0`
The `0.15.0` was a massive release for `hashes` in terms of API clean
up and also adding new HKDF support. ElligatorSwift in combination with
HKDF are the primitives required for encrypted message passing on
bitcoin, so exporting these in tandum would save an additional
`hashes` dependency.

I skip to `0.16.0` because `hashes` allows for `hex-conservative` as an
optional dependency, and no listed changes elsewhere.