Enable hashes alloc feature
The feature enabling is borked at the moment. "alloc" should enabled
"alloc" ifn optional `hashes` dependency.

The bug is not showing up, my best guess is because in `hashes v0.14`
we probably had feature gating nested in macros called across crate
boundary's and were either getting or not getting code that we thought
we should or shouldn't - what ever the case may be.

<insert rand about macros here>

In order to upgrade to `hashes 0.16.0` we need to fix this.