secret: stop using "hashes" dependency for secret debug output
This is the only place we use a sha256 hash. We were using the 'hashes'
crate which is a whole extra dependency just to bring in a hash function
that already exists in libsecp256k1.

libsecp256k1 doesn't expose sha256 directly, but they do expose default
hash functions for EDCH, ECDSA and Schnorr, which are just sha2s with
extra steps. Since we are not worried about having a specific canonical
hash, just one that won't change for different keys, do this.

(This change does mean that users will get different hashes when they
upgrade this library; but they really shouldn't be persisting Debug
output and expecting it to be consistent. I don't think this is a
meaningful breaking change.)