remove use of 'hashes' from public API
We had an impl of the (deprecated) ThirtyTwoByte hash trait, which we no
longer use; we delete this. Also remove the 'hashes' feature gate from a
unit test that no longer needs it.

There is also a `pub extern crate hashes` in src/lib.rs which we should
remove here, but it's used in some doctests so we'll remove it in the
next commit.