ecdh: remove use of `hashes` in example code and tests
We have a curious test where we confirm that we can match upstream's code
with direct calls to sha256. This is nice, but completely unnecessary as
a unit test.

Secondly, we have example code where we demonstrate how to use bitcoin_hashes
to write a custom ECDH nonce function. This is valuable, but not reason by
itself to have an extra dependency, so we stick `ignore` on it to prevent
the Rust compiler from attempting to compile it.