remove use of "hashes" from examples and doccomments
This is a little annoying; the library has a Message type for
ECDSA signing which is supposed to be a secure hash digest,
but we provide no methods for producing such a digest. So
our examples are necessarily a little wonky.

Before this commit we would use `bitcoin_hashes` to hash some
arbitrary data. This isn't really wise (signing unstructured
data or using hashes that aren't domain-separated) but it was
shaped similarly to proper usage.

Now we just hardcode the "output" of hash functions and convert
that. I added some comments to help.