secp-sys: update raw pointer deref syntax in fuzz version of the crate
Rust 1.88.0 introduced a warning, and 1.89.0 made it an error, to dereference
a raw pointer in a way where it is implicitly turned into a reference. This
is probably a good lint but breaking compilation like this is bullshit.

There are many instances of this in our fuzz code (none in the real code).
We just need to change the implicit & and &muts to explicit ones.

We will need to backport some change if we want our CI to keep working on
old versions. Probably just whitelisting the lint for old versions of the
library is fine.