Put back in deprecated context methods
During the removal of the context from the public API I brashly
deleted a bunch of impls on the `Secp256k1` type when we could have
just deprecated them. This means that when upgrading downstream crates
the upgrade is way more painful than it is if we deprecate.

Add back in the context impls. To do this I copied from `v0.31.1`. I'm
not totally sure that there wasn't changes to this code in `master`
but since we are explicitly doing this to help upgrade then it seems
reasonable to do it like this.

Add deprecated since 0.32 attributes because we are going to push this
as a `0.32.0-beta.1` release.

Sorry for the pain in reviewing this, I'm not entirely sure how you
are supposed to convince yourself that the code is exactly as it was?